% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in fars file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{character vectors, containing file names or paths.  In addiition
to csv files, the function will take compressed csv files with the file extensions
\code{.gz .bz2 .xz .zip}}
}
\value{
a data frame containing the parsed file data.
}
\description{
This is a simple function that uses the `readr` package to read in a csv file
and then uses `dplyr` to convert the data from the csv into a data frame.
}
\examples{
\dontrun{fars_read('myFile.csv')}
}
