% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple fars files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of years in YYYY format corresponding to files to be parsed.}
}
\value{
a data frame containing data from the files that have been parsed.
}
\description{
This function attempts to read in data files given a vector of years.
}
\examples{
\dontrun{fars_read_years(c(2013, 2014, 2015))}
}
