% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize fars data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of years from which to extract data from flat files.}
}
\value{
a list with years as columns and counts of accidents as observations.
}
\description{
Summarizes fars data by creating a table with years as columns and counts of
observations for each month in each row.
}
\examples{
\dontrun{fars_read_years(c(2013, 2014, 2015))}
}
