% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a non-negative number that can be coerced into an integer}
}
\value{
a file string reading \code{"accident_[YEAR].csv.bz2"}
}
\description{
This function creates a file string for an accident file based on the inputted year
}
\examples{
\dontrun{ make_filename("2009")}
}
